backend/requirements.txt
# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.0.0
python-dotenv>=1.0.0

# AI/LangChain
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# Computer Vision
opencv-python>=4.9.0
mediapipe>=0.10.0
numpy>=1.23.0

# Utilities
Pillow>=10.0.0
